backend/requirements.txt
fastapi
uvicorn[standard]
pydantic

# Graph / workflow engine
langgraph
langsmith

# HTTP client for tool adapters
httpx

# Object storage (MinIO / S3-compatible)
minio

# LLM client
openai
python-dotenv

# You can add other LLM SDKs here during the hackathon, e.g.:
# anthropic

# Testing
pytest
pytest-asyncio
